+++ DatabaseMicroservice/src/main/java/com/licenta/databasemicroservice/business/service/UserDetailsService.java
@@ -2,12 +2,10 @@ package com.licenta.databasemicroservice.business.service;
 
 import com.licenta.databasemicroservice.business.interfaces.ICityService;
 import com.licenta.databasemicroservice.business.interfaces.IUserDetailsService;
-import com.licenta.databasemicroservice.business.interfaces.IUserService;
 import com.licenta.databasemicroservice.business.model.userdetails.SaveUserDetailsRequest;
 import com.licenta.databasemicroservice.business.model.userdetails.UserDetailsResponse;
 import com.licenta.databasemicroservice.business.util.mapper.UserDetailsMapper;
 import com.licenta.databasemicroservice.persistence.entity.City;
-import com.licenta.databasemicroservice.persistence.entity.User;
 import com.licenta.databasemicroservice.persistence.entity.UserDetails;
 import com.licenta.databasemicroservice.persistence.repository.UserDetailsRepository;
 import org.mapstruct.factory.Mappers;
@@ -26,8 +24,7 @@ import java.io.IOException;
 @Service
 public class UserDetailsService implements IUserDetailsService {
 
-    @Autowired
-    private IUserService userService;
+
     @Autowired
     private ICityService cityService;
     @Autowired
@@ -48,11 +45,13 @@ public class UserDetailsService implements IUserDetailsService {
     public UserDetailsResponse saveUserDetails(Long userId, SaveUserDetailsRequest request) {
         UserDetails userDetails = userDetailsMapper.toModel(request);
 
-        User user = userService.getUserOrElseThrowException(userId);
-        City city = cityService.getCityOrElseThrowException(request.getCityId());
 
-        userDetails.setUserId(user.getId());
-        userDetails.setCity(city);
+        userDetails.setUserId(userId);
+
+        if(request.getCityId() != null) {
+            City city = cityService.getCityOrElseThrowException(request.getCityId());
+            userDetails.setCityId(city.getId());
+        }
 
         if (userDetailsRepository.findById(userId).isPresent()) {
             userDetails.setUserId(userId);
@@ -65,9 +64,6 @@ public class UserDetailsService implements IUserDetailsService {
 
     @Override
     public UserDetailsResponse getUserDetails(Long userId) {
-
-        userService.getUserOrElseThrowException(userId);
-
         UserDetails  userDetails = userDetailsRepository.findById(userId).orElse(null);
 
         if(userDetails != null)
@@ -92,11 +88,13 @@ public class UserDetailsService implements IUserDetailsService {
 
         String extension = mimeContentService.getExtensionFromMimeType(type);
 
-        String[] arr = userDetails.getProfilePictureUrl().split("/");
-        String currentImagePath = arr[arr.length - 1];
-        String oldPath = usersImagesPath + "/" + currentImagePath;
-        File file = new File(oldPath);
-        file.delete();
+        if(userDetails.getProfilePictureUrl() != null) {
+            String[] arr = userDetails.getProfilePictureUrl().split("/");
+            String currentImagePath = arr[arr.length - 1];
+            String oldPath = usersImagesPath + "/" + currentImagePath;
+            File file = new File(oldPath);
+            file.delete();
+        }
 
         String imageName = userId + "." + extension;
         String path = usersImagesPath + "/" + imageName;
@@ -104,9 +102,15 @@ public class UserDetailsService implements IUserDetailsService {
             fos.write(imageBytes);
             fos.flush();
         }
+        System.out.println(imageName);
 
         path = urlService.getBaseUrl() + "/images/users/" + imageName;
         userDetails.setProfilePictureUrl(path);
         userDetailsRepository.save(userDetails);
+    }
+
+    @Override
+    public void deleteUserDetails(Long userId) {
+        userDetailsRepository.deleteById(userId);
     }
 }
